refactor(sim): share build-from-template primitives across Astral + SCIEX
The schedule->tables transform and the two duck-typed Bruker-reference
stand-ins (helper-handle + TDFWriter stub) were defined in
astral_acquisition but cross-imported by the SCIEX builder under their
private _Astral* names — misleading now that they are instrument-neutral.

Move them to a new public module template_acquisition_common with neutral
names (build_frame_tables_from_schedule, build_synthetic_scan_table,
TemplateHelperHandle, TemplateTdfWriterStub). astral_acquisition re-exports
the originals as back-compat aliases (build_astral_frame_tables /
_AstralHelperHandle / _AstralTdfWriterStub) so existing callers and the
P6e test are untouched; SCIEX imports the canonical module directly.

Pure relocation — no behavior change. New test locks the alias contract;
the existing Astral test still passes through the alias.
