+++ packages/imspy-simulation/src/imspy_simulation/timsim/jobs/astral_acquisition.py
@@ -13,140 +13,27 @@ testable on its own; the acquisition builder consumes its output.
 
 from __future__ import annotations
 
-from typing import List, Optional, Sequence, Tuple
+from typing import Optional
 
 import numpy as np
-import pandas as pd
 
-# A schedule row: (scan, rt_s, ms_level, center_mz|None, width_mz|None, ce|None)
-ScheduleRow = Tuple[int, float, int, Optional[float], Optional[float], Optional[float]]
+# The schedule->tables transform and the duck-typed Bruker-reference stand-ins are
+# instrument-neutral (the SCIEX build-from-.wiff path reuses them verbatim), so they
+# live in the shared module. Imported here under both the canonical neutral names and
+# the original ``*astral*`` / ``_Astral*`` names so existing call sites and tests keep
+# working.
+from .template_acquisition_common import (
+    ScheduleRow,
+    TemplateHelperHandle,
+    TemplateTdfWriterStub,
+    build_frame_tables_from_schedule,
+    build_synthetic_scan_table,
+)
 
-
-def build_synthetic_scan_table(
-    num_scans: int = 451,
-    im_lower: float = 0.6,
-    im_upper: float = 1.6,
-) -> pd.DataFrame:
-    """A synthetic mobility (scan) grid for an Astral run — NO Bruker reference.
-
-    Astral has no ion mobility, and the render marginalises the mobility axis away
-    (P6c/P6e), so the actual 1/K0 values do not affect Astral output — but the
-    simulation pipeline still projects ion-mobility distributions onto a scan grid,
-    so it needs SOME grid. We provide a plausible descending 1/K0 grid (the Bruker
-    `scans` table is descending scan index with ascending mobility) without reading
-    a reference `.d`. This is the key decoupling primitive for the lean Astral
-    acquisition path (option b): it removes the `TDFWriter.helper_handle` dependency.
-    """
-    if num_scans < 1:
-        raise ValueError("num_scans must be >= 1")
-    if not (im_lower < im_upper):
-        raise ValueError("im_lower must be < im_upper")
-    # Descending scan index (Bruker convention), ascending inverse mobility.
-    scans = np.arange(num_scans, dtype=np.int64)[::-1]
-    mobilities = np.linspace(im_lower, im_upper, num_scans, dtype=np.float64)
-    return pd.DataFrame({"scan": scans, "mobility": mobilities})
-
-
-def build_astral_frame_tables(
-    schedule: Sequence[ScheduleRow],
-    *,
-    num_scans: int = 1,
-    ce_decimals: int = 2,
-) -> Tuple[pd.DataFrame, pd.DataFrame, pd.DataFrame, List[int]]:
-    """Build the Astral acquisition tables from a template scan schedule.
-
-    Each template scan becomes one frame (``frame_id`` 1..N in scan order), timed
-    at its REAL retention time. MS1 scans are precursor frames (``ms_type`` 0); MS2
-    scans are fragment frames (``ms_type`` 9) carrying the template's isolation
-    window + collision energy. Distinct (center, width, CE) windows are assigned
-    window-group ids; because Astral has no ion mobility, each window spans the
-    FULL scan range ``[0, num_scans-1]`` so the (mobility-indexed) transmission
-    degenerates to pure m/z gating.
-
-    Returns ``(frame_table, dia_ms_ms_windows, frames_to_window_groups,
-    frame_to_template_scan)`` where the last is a list mapping ``frame_id-1`` ->
-    the template scan number (so the writer can author each frame into its slot).
-    """
-    if not schedule:
-        raise ValueError("empty template schedule")
-
-    frames = []
-    fragment_rows = []  # (frame_id, window_key)
-    frame_to_template_scan: List[int] = []
-    # Deterministic window-group assignment: first occurrence order.
-    window_key_to_group: dict[Tuple[float, float, float], int] = {}
-    window_group_meta: dict[int, Tuple[float, float, float]] = {}
-
-    n_ms1 = 0
-    for idx, row in enumerate(schedule):
-        scan, rt, ms_level, center, width, ce = row
-        frame_id = idx + 1
-        frame_to_template_scan.append(int(scan))
-        is_ms1 = ms_level <= 1
-        ms_type = 0 if is_ms1 else 9
-        frames.append({"frame_id": frame_id, "time": float(rt), "ms_type": ms_type})
-        if is_ms1:
-            n_ms1 += 1
-            continue
-        # MS2: resolve / assign its window group.
-        if center is None or width is None:
-            raise ValueError(f"MS2 scan {scan} (frame {frame_id}) has no isolation window")
-        ce_val = round(float(ce), ce_decimals) if ce is not None else 0.0
-        key = (round(float(center), 4), round(float(width), 4), ce_val)
-        group = window_key_to_group.get(key)
-        if group is None:
-            group = len(window_key_to_group) + 1  # 1-based window-group ids
-            window_key_to_group[key] = group
-            window_group_meta[group] = key
-        fragment_rows.append({"frame": frame_id, "window_group": group})
-
-    if n_ms1 == 0:
-        raise ValueError("template schedule has no MS1 scans")
-    if not fragment_rows:
-        raise ValueError("template schedule has no MS2 scans")
-
-    frame_table = pd.DataFrame(frames)
-    frames_to_window_groups = pd.DataFrame(fragment_rows)
-
-    # One dia_ms_ms_windows row per window group. Astral has no mobility, so the
-    # window spans the whole scan range (mobility transmission -> pure m/z gating).
-    win_rows = []
-    for group, (center, width, ce_val) in sorted(window_group_meta.items()):
-        win_rows.append(
-            {
-                "window_group": group,
-                "scan_start": 0,
-                "scan_end": int(max(num_scans - 1, 0)),
-                "isolation_mz": center,
-                "isolation_width": width,
-                "collision_energy": ce_val,
-            }
-        )
-    dia_ms_ms_windows = pd.DataFrame(win_rows)
-
-    return frame_table, dia_ms_ms_windows, frames_to_window_groups, frame_to_template_scan
-
-
-class _AstralHelperHandle:
-    """Minimal stand-in for a Bruker reference's ``helper_handle`` — exposes only
-    the m/z + ion-mobility ranges the distribution jobs read. Astral has no IMS;
-    the mobility range is synthetic (marginalised away at render)."""
-
-    def __init__(self, mz_lower, mz_upper, im_lower, im_upper, num_scans):
-        self.mz_lower = mz_lower
-        self.mz_upper = mz_upper
-        self.im_lower = im_lower
-        self.im_upper = im_upper
-        self.num_scans = num_scans
-
-
-class _AstralTdfWriterStub:
-    """Stand-in for the Bruker ``TDFWriter``: the Astral path writes a Thermo
-    ``.raw`` (not a ``.d``), so only the reference ranges are needed during
-    simulation, never the binary writer."""
-
-    def __init__(self, helper_handle):
-        self.helper_handle = helper_handle
+# Backward-compatible aliases (the function/classes moved to template_acquisition_common).
+build_astral_frame_tables = build_frame_tables_from_schedule
+_AstralHelperHandle = TemplateHelperHandle
+_AstralTdfWriterStub = TemplateTdfWriterStub
 
 
 class AstralAcquisitionBuilder:
@@ -211,11 +98,11 @@ class AstralAcquisitionBuilder:
             )
 
         # Round (and window-group) collision energies at the configured precision —
-        # do it ONCE inside build_astral_frame_tables so grouping and the stored CE
-        # use the same values (no premature truncation). round_collision_energy=False
+        # do it ONCE inside build_frame_tables_from_schedule so grouping and the stored
+        # CE use the same values (no premature truncation). round_collision_energy=False
         # keeps full precision (grouped at 6 dp to avoid float over-splitting).
         ce_decimals = collision_energy_decimals if round_collision_energy else 6
-        ft, win, f2g, f2scan = build_astral_frame_tables(
+        ft, win, f2g, f2scan = build_frame_tables_from_schedule(
             schedule, num_scans=num_scans, ce_decimals=ce_decimals
         )
         # Optional manual override: force a single NCE across all windows. Off by
@@ -238,8 +125,8 @@ class AstralAcquisitionBuilder:
             mz_lower = float(win["isolation_mz"].min() - win["isolation_width"].max())
         if mz_upper is None:
             mz_upper = float(win["isolation_mz"].max() + win["isolation_width"].max())
-        self.tdf_writer = _AstralTdfWriterStub(
-            _AstralHelperHandle(mz_lower, mz_upper, im_lower, im_upper, num_scans)
+        self.tdf_writer = TemplateTdfWriterStub(
+            TemplateHelperHandle(mz_lower, mz_upper, im_lower, im_upper, num_scans)
         )
 
         for name, tbl in (
